Add management command to generate areasearch form
